more 32

+++ spit/jobType.h
@@ -49,7 +49,7 @@ int jobRunThreads(jobType *job, const int num, char *filePrefix,
                    unsigned short seed, FILE *savePositions, diskStatType *d, const double timeperline,
                    const double ignorefirst, const size_t verify,
                    char *mysqloptions, char *mysqloptions2, char *commandstring, char *numaBinding,
-                   const int performDiscard, resultType *result, size_t ramBytesForPositions, size_t notexclusive,
+                   const int performDiscard, resultType *result, long long ramBytesForPositions, size_t notexclusive,
                    const size_t showdate, FILE *loadpos, const double exitTimeout, const size_t showASCII, const size_t exitOnErrors);
 
 void jobMultiply(jobType *retjobs, jobType *job, deviceDetails *deviceList, size_t deviceCount);
